WIP - refactoring navbar code
#331
