+++ waltz-ng/client/navbar/directives/nav-search-results.html
@@ -1,4 +1,4 @@
-<div style="margin: 10px;">
+<div class="waltz-nav-search-results">
     <h3 style="border-bottom: 1px solid #ccc">
         Search Results: {{query}}
         <small class="pull-right">
