+++ waltz-ng/client/navbar/directives/nav-search-results.js
