+++ waltz-ng/client/navbar/directives/navbar-recently-viewed.html
@@ -1,0 +1,32 @@
+<li class="dropdown" uib-dropdown>
+    <a class="clickable"
+       ng-click="ctrl.refreshHistory()"
+       uib-dropdown-toggle>
+        <waltz-icon name="history"></waltz-icon>
+        Recently Viewed
+        <b class="caret"></b>
+    </a>
+    <ul class="dropdown-menu"
+        uib-dropdown-menu
+        style="width: 250px; margin-right:8px"
+        role="menu"
+        class="no-overflow">
+        <li style="margin: 4px"
+            class="no-overflow"
+            ng-repeat="item in ctrl.history">
+            <a ui-sref="{{ ::item.state }} ({{ ::item.stateParams }})">
+                <waltz-icon class='text-muted small'
+                            fixed-width="true"
+                            name="{{ item.kind | toIconName:'entity' }}"></waltz-icon>
+                <span ng-bind="::item.name"></span>
+            </a>
+        </li>
+
+        <li class="text-muted"
+            ng-if="ctrl.history.length == 0">
+                        <span style="margin: 4px">
+                            Recently viewed items will be added as you navigate through Waltz
+                        </span>
+        </li>
+    </ul>
+</li>
