+++ waltz-ng/client/navbar/index.js
@@ -1,4 +1,3 @@
-
 /*
  *  Waltz
  * Copyright (c) David Watkins. All rights reserved.
@@ -11,5 +10,12 @@
  *
  */
 
+export default (module) => {
 
-export default (module) => module.directive('waltzNavbar', require('./navbar'));
+    module
+        .directive('waltzNavbar', require('./directives/navbar'))
+        .directive('waltzNavbarRecentlyViewed', require('./directives/navbar-recently-viewed'))
+        .directive('waltzNavSearchResults', require('./directives/nav-search-results'));
+
+    module.component('waltzNavbarSearchForm', require('./components/navbar-search-form'));
+};
