+++ waltz-ng/client/navbar/styles.scss
@@ -1,0 +1,1 @@
+@import "./directives/nav-search-results";
