+++ waltz-ng/client/widgets/index.js
@@ -22,7 +22,6 @@ export default (module) => {
     module.directive('waltzKeywordList', require('./keyword-list'));
     module.directive('waltzLoadingNotification', require('./loading-notification'));
     module.directive('waltzMailTo', require('./mail-to'));
-    module.directive('waltzNavSearchResults', require('./nav-search-results'));
     module.directive('waltzOverlayPanel', require('./overlay-panel'));
     module.directive('waltzPageHeader', require('./page-header'));
     module.directive('waltzPhoneLink', require('./phone-link'));
