+++ waltz-ng/style/_waltz.scss
@@ -26,6 +26,7 @@
 @import "../client/data-type-usage/directives/app-data-type-usage-editor";
 @import "../client/entity-statistics/directives/entity-statistic";
 @import "../client/entity-statistics/components/entity-statistic-summary-card";
+@import "../client/navbar/styles";
 @import "../client/ratings/directives/rating-group/rating-group";
 @import "../client/playpen/playpen";
 @import "../client/ratings/directives/rating-brush-select/style";
